requirements.txt
flask==3.0.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
tensorflow==2.15.0
pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4